Change std::min uses to operate on uint64_t since it might lead to overflow issues

+++ bitreader/include/bitreader/bitreader.hpp
@@ -159,7 +159,7 @@ namespace brcpp {
         //----------------------------------------------------------------------
         void _next(internal_state& state) const
         {
-            size_t available = std::min<size_t>(
+            size_t available = std::min<uint64_t>(
                     sizeof(state.buffer),
                     state.source->available());
